requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
psycopg2-binary==2.9.9

# Vector DB
qdrant-client==1.9.1

# AI / Embeddings
google-generativeai==0.7.2

# PDF parsing
pdfplumber==0.11.0

# HTTP / scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# Task queue
celery==5.4.0
redis==5.0.6

# Auth
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9

# LiveKit voice agent
livekit-agents
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-elevenlabs
livekit-plugins-deepgram
